Standardize on big-endian terminology in byteutils
Not sure about this. Big endian is easier to model mentally, and most people are used to working in big endian because pretty much everything except x86 processors use it. However, the rest of the codebase uses little endian, and an abrupt shift in one module might create more confusion than it solves.